pyproject.toml
[project]
name = "restic-configurator-py"
version = "0.1.0"
dependencies = [
    "python-dotenv >= 1.0.0",
]